Implement role checking in delete_access

+++ src/api/ecdar_api.rs
@@ -605,6 +605,46 @@ impl EcdarApi for ConcreteEcdarApi {
     ) -> Result<Response<()>, Status> {
         let message = request.get_ref().clone();
 
+        let uid = request
+            .uid()
+            .ok_or(Status::internal("Could not get uid from request metadata"))?;
+
+        let user_access = self
+            .contexts
+            .access_context
+            .get_by_id(message.id)
+            .await
+            .map_err(|err| Status::new(Code::Internal, err.to_string()))?
+            .ok_or_else(|| {
+                Status::new(
+                    Code::NotFound,
+                    "No access entity found for user".to_string(),
+                )
+            })?;
+
+        check_editor_role_helper(
+            Arc::clone(&self.contexts.access_context),
+            uid,
+            user_access.model_id,
+        )
+        .await?;
+
+        let model = self
+            .contexts
+            .model_context
+            .get_by_id(user_access.model_id)
+            .await
+            .map_err(|err| Status::new(Code::Internal, err.to_string()))?
+            .ok_or_else(|| Status::new(Code::NotFound, "No model found for access".to_string()))?;
+
+        // Check that the requester is not trying to delete the owner's access
+        if model.owner_id == message.id {
+            return Err(Status::new(
+                Code::PermissionDenied,
+                "You cannot delete the access entity for this user",
+            ));
+        }
+
         match self.contexts.access_context.delete(message.id).await {
             Ok(_) => Ok(Response::new(())),
             Err(error) => match error {
